backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
pgvector==0.3.6

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Environment Variables
python-decouple==3.8

# Password Hashing
argon2-cffi==23.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Report Generation
reportlab==4.1.0
openpyxl==3.1.2

# Machine Learning & Predictive Analytics
scikit-learn==1.4.0
statsmodels==0.14.1

# AI/LLM Providers
anthropic>=0.40.0
openai>=1.12.0

# FastAPI Streaming Service
fastapi>=0.110.0
uvicorn>=0.27.0
pyjwt>=2.8.0

# API Documentation
drf-spectacular==0.27.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Celery (Background Tasks)
celery==5.3.4
redis==5.0.1

# Rate Limiting
django-ratelimit==4.1.0

# Security - File Type Validation (validates file content, not just extension)
python-magic==0.4.27

# Security - Encrypted Fields (optional, enable if FIELD_ENCRYPTION_KEY is set)
# django-encrypted-model-fields==0.6.5
# cryptography>=41.0.0

# Development
django-extensions==3.2.3
ipython==8.20.0

# Testing
pytest==8.3.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.24.0
factory-boy==3.3.0
faker==22.0.0
responses==0.24.1
pytest-mock==3.12.0
